+++ backend/src/main/java/com/wooteco/nolto/feed/application/FeedService.java
@@ -2,7 +2,7 @@ package com.wooteco.nolto.feed.application;
 
 import com.wooteco.nolto.NotFoundException;
 import com.wooteco.nolto.feed.domain.Feed;
-import com.wooteco.nolto.feed.domain.FeedRepository;
+import com.wooteco.nolto.feed.domain.repository.FeedRepository;
 import com.wooteco.nolto.feed.domain.Feeds;
 import com.wooteco.nolto.feed.domain.FilterStrategy;
 import com.wooteco.nolto.feed.ui.dto.FeedCardResponse;
