+++ backend/src/main/java/com/wooteco/nolto/feed/domain/repository/FeedRepository.java
@@ -1,5 +1,6 @@
-package com.wooteco.nolto.feed.domain;
+package com.wooteco.nolto.feed.domain.repository;
 
+import com.wooteco.nolto.feed.domain.Feed;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
