+++ backend/src/main/java/com/wooteco/nolto/feed/domain/repository/FeedTechRepository.java
@@ -1,5 +1,6 @@
-package com.wooteco.nolto.feed.domain;
+package com.wooteco.nolto.feed.domain.repository;
 
+import com.wooteco.nolto.feed.domain.FeedTech;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FeedTechRepository extends JpaRepository<FeedTech, Long> {
